# Optimize text mimerenderer: ansi vs autolink 