--- a/packages/rendermime/test/factories.spec.ts
+++ b/packages/rendermime/test/factories.spec.ts
@@ -86,7 +86,15 @@ describe('rendermime/factories', () => {
         [
           '\x1b[48;2;185;0;129mwww.example.\x1b[0m\x1b[48;2;113;0;119mcom\x1b[0m',
           '<pre><a href="https://www.example.com" rel="noopener" target="_blank"><span style="background-color:rgb(185,0,129)">www.example.</span><span style="background-color:rgb(113,0,119)">com</span></a></pre>'
-        ]
+        ],
+        [
+          'Prefix \x1b[48;2;185;0;129m spacer www.example.\x1b[0m\x1b[48;2;113;0;119mcom\x1b[0m',
+          '<pre>Prefix <span style="background-color:rgb(185,0,129)"> spacer </span><a href="https://www.example.com" rel="noopener" target="_blank"><span style=\"background-color:rgb(185,0,129)\">www.example.</span><span style="background-color:rgb(113,0,119)">com</span></a></pre>'
+        ],
+        [
+          'Prefix www.example.\x1b[0m\x1b[48;2;113;0;119mcom postfix\x1b[0m',
+          '<pre>Prefix <a href="https://www.example.com" rel="noopener" target="_blank">www.example.<span style="background-color:rgb(113,0,119)">com</span></a><span style="background-color:rgb(113,0,119)"> postfix</span></pre>'
+        ],
       ])(
         'should output the correct HTML with ansi colors',
         async (source, expected) => {
@@ -165,18 +173,18 @@ describe('rendermime/factories', () => {
           })
         );
       });
-    });
 
-    it('should autolink multiple URLs', async () => {
-      const source = 'www.example.com\nwww.python.org';
-      const expected =
-        '<pre><a href="https://www.example.com" rel="noopener" target="_blank">www.example.com</a>\n<a href="https://www.python.org" rel="noopener" target="_blank">www.python.org</a></pre>';
-      const f = textRendererFactory;
-      const mimeType = 'text/plain';
-      const model = createModel(mimeType, source);
-      const w = f.createRenderer({ mimeType, ...defaultOptions });
-      await w.renderModel(model);
-      expect(w.node.innerHTML).toBe(expected);
+      it('should autolink multiple URLs', async () => {
+        const source = 'www.example.com\nwww.python.org';
+        const expected =
+          '<pre><a href="https://www.example.com" rel="noopener" target="_blank">www.example.com</a>\n<a href="https://www.python.org" rel="noopener" target="_blank">www.python.org</a></pre>';
+        const f = textRendererFactory;
+        const mimeType = 'text/plain';
+        const model = createModel(mimeType, source);
+        const w = f.createRenderer({ mimeType, ...defaultOptions });
+        await w.renderModel(model);
+        expect(w.node.innerHTML).toBe(expected);
+      });
     });
   });