## References

Fixes JS tests on Windows by removing some backslashes that were added in #10301.

Optimizes ansi color coding vs auto-linking in plain text rendermime. #11272

## Code changes

This optimizes the rendering for cases where there are many more elements of one type than the other. E.g. for a large syntax highlighted source file. In this case there can be many (tens of) thousands of span elements, but likely only a few links. In this case, multiple span elements will all correspond to a single text node in the autolink output. In these cases, we can simply use the nodes from the span element list, without splitting / cloning any nodes. This greatly reduces the time spent.

## User-facing changes

UI hangs less. For an extremely symptomatic file in the inspector, time spent is reduced from ~40s to ~4s.

Old, 41s:
![renderText-old-41s](https://user-images.githubusercontent.com/510760/194617139-8edbc4c0-4412-4deb-bfb5-38e3c77106cd.png)

New, 4s:
![renderText-new-3s](https://user-images.githubusercontent.com/510760/194617148-bd4f7874-acf4-486a-a968-8283e140d457.png)

## Backwards-incompatible changes

Should only change internal details. Ideally this would be backported to 3.x , but then we would likely need to rewrite using lumino iterators.