@vidartf 2022-10-14
@fcollonval it should be now, but let's see if CI has anything to say on the matter 😆