@fcollonval 2022-10-14
@meeseeksdev please backport to 3.5.x